requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.16.0,<4.0
djangorestframework-simplejwt>=5.5.0,<6.0
django-cors-headers>=4.9.0,<5.0
